requirements.txt
h5py==2.9.0 
imageio==2.9.0 
Keras==2.4.3 
matplotlib==3.3.3 
nibabel==3.2.1 
numpy==1.19.4 
scikit-image 
scikit-learn 
scipy==1.5.4 
sklearn 
tensorflow-gpu==1.15

